+++ odbcsv/src/main.rs
@@ -293,11 +293,16 @@ fn insert(environment: &Environment, insert_opt: &InsertOpt) -> Result<(), Error
     // Allocate buffer
     let mut buffer = TextRowSet::new(*batch_size, (0..headline.len()).map(|_| 0));
 
+    // Used to log batch number
+    let mut num_batch = 0;
+
     for try_record in reader.into_byte_records() {
         if buffer.num_rows() == *batch_size as usize {
+            num_batch += 1;
             // Batch is full. We need to send it to the data base and clear it, before we insert
             // more rows into it.
             statement.execute(&buffer)?;
+            info!("Insert batch {} with {} rows into DB.", num_batch, batch_size);
             buffer.clear();
         }
 
@@ -312,6 +317,7 @@ fn insert(environment: &Environment, insert_opt: &InsertOpt) -> Result<(), Error
     // Insert the remainder of the buffer to the database. If buffer is empty nothing will be
     // executed.
     statement.execute(&buffer)?;
+    info!("Insert last batch with {} rows into DB.", batch_size);
 
     Ok(())
 }